+++ src/dune/main.ml
@@ -118,7 +118,7 @@ let auto_concurrency =
               | None -> loop rest
               | Some prog -> (
                 let* result =
-                  Process.run_capture (Accept Predicate_lang.Ast.any) prog args
+                  Process.run_capture (Accept Predicate_lang.any) prog args
                     ~env:Env.initial
                     ~stderr_to:(Process.Io.file Config.dev_null Process.Io.Out)
                 in