+++ src/dune/process.ml
@@ -4,14 +4,14 @@ open Fiber.O
 
 type ('a, 'b) failure_mode =
   | Strict : ('a, 'a) failure_mode
-  | Accept : int Predicate_lang.Ast.t -> ('a, ('a, int) result) failure_mode
+  | Accept : int Predicate_lang.t -> ('a, ('a, int) result) failure_mode
 
 let accepted_codes : type a b. (a, b) failure_mode -> int -> bool = function
   | Strict -> Int.equal 0
   | Accept exit_codes ->
     fun i ->
-      Predicate_lang.Ast.exec exit_codes
-        ~standard:(Predicate_lang.Ast.Element 0) (Int.equal i)
+      Predicate_lang.exec exit_codes ~standard:(Predicate_lang.Element 0)
+        (Int.equal i)
 
 let map_result :
     type a b. (a, b) failure_mode -> int Fiber.t -> f:(unit -> a) -> b Fiber.t