+++ CHANGES.md
@@ -146,13 +146,17 @@
   self-contained bytecode executables whenever this options is available
   (OCaml version >= 4.10) (#2692, @nojb)
 
-- Add action `(with-exit-codes <pred> <action>)` to specify the set of
+- Add action `(with-accepted-exit-codes <pred> <action>)` to specify the set of
   successful exit codes of `<action>`. `<pred>` is specified using the predicate
   language. (#2699, @nojb)
 
 - Do not setup rules for disabled libraries (#2491, fixes #2272, @bobot)
 
 - Configurator: filter out empty flags from `pkg-config` (#2716, @AltGr)
+
+- `no_keep_locs` is a no-op for projects that use `lang dune` older than 2.0. In
+  projects where the language is at least `2.0`, the field is now forbidden.
+  (#2752, fixes #2747, @rgrinberg)
 
 1.11.4 (09/10/2019)
 -------------------